Fixed the wrong datatype of stats_avg[8][2]
	modified:   sw/or_data_types.h
		 The datatype of stats_avg[8][2] should be 'long double' instead of 'double'

	modified:   sw/or_netfpga.c
		 Now rs->stats_avg[i][x] can accept the values correctly

+++ sw/or_data_types.h
@@ -280,7 +280,7 @@ struct router_state {
 	pthread_mutex_t* stats_mutex;
 	struct timeval stats_last_time;
 	uint32_t stats_last[8][4];
-	double stats_avg[8][2];
+	long double stats_avg[8][2];
 
 	pthread_mutex_t* local_ip_filter_list_mutex;
 	node* local_ip_filter_list;
